Moving code to utils to be used from another projects
